apps/architect/requirements.txt
# UI & Software Window
PyGObject
nicegui
pywebview

# Agents & Logic
langgraph
langchain-ollama
chromadb
pydantic-settings

# Monitoring & Tracing
arize-phoenix
opentelemetry-sdk
opentelemetry-exporter-otlp
openinference-instrumentation-langchain

# Testing
pytest
pytest-env
pytest-bdd
pytest-asyncio 
nest-asyncio
pytest-dotenv

# Data Analysis
numpy
pandas
scipy